[ADD-UNIT-TEST]: Move test configuration to configuration package. Fix gradle task to download the grpc-csharp compiler for linux.
