+++ backend/src/test/java/com/example/demo/configuration/GrpcIntegrationTestConfiguration.java
@@ -1,4 +1,4 @@
-package com.example.demo.grpc;
+package com.example.demo.configuration;
 
 import net.devh.boot.grpc.client.autoconfigure.GrpcClientAutoConfiguration;
 import net.devh.boot.grpc.server.autoconfigure.GrpcServerAutoConfiguration;
