+++ protobuf/build.gradle
@@ -42,7 +42,7 @@ def downloadGrpcToCsharpCompiler(String path, osFamily) {
         if ("${osFamily}".contains('windows')) {
             execute(['cmd', '/c', 'csharp-grpc-downloader.bat'])
         } else if ("${osFamily}".contains('linux')) {
-            execute(['sh', '-c', 'csharp-grpc-downloader'])
+            execute(['sh', '-c', 'csharp-grpc-downloader.sh'])
         }
     }
 }
